Extract interface for OrgsDao/OrgsDaoImpl
