+++ console/src/test/java/org/georchestra/console/integration/OrgsIT.java
@@ -22,6 +22,7 @@ import static org.hamcrest.Matchers.stringContainsInOrder;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -76,7 +77,7 @@ public class OrgsIT extends ConsoleIntegrationTest {
 
         create(orgName);
 
-        support.perform(get("/private/orgs/" + orgName)).andExpect(jsonPath("$.description").value(""));
+        support.perform(get("/private/orgs/" + orgName)).andDo(print()).andExpect(jsonPath("$.description").value(""));
     }
 
     @WithMockUser(username = "admin", roles = "SUPERUSER")
