+++ ldap-account-management/src/main/java/org/georchestra/ds/users/AccountDao.java
@@ -167,5 +167,7 @@ public interface AccountDao {
 
     boolean hasUserLoginChanged(Account account, Account modified);
 
+    // TODO: this is not API, but LDAP specific implementation detail. Make it just
+    // a static method in AccountDaoImpl
     String buildFullUserDn(Account account);
 }
