+++ ldap-account-management/src/test/resources/testApplicationContext.xml
@@ -57,7 +57,7 @@
     <property name="roleSearchBaseDN" value="${ldapRolesRdn}"/>
   </bean>
 
-  <bean id="orgsDao" class="org.georchestra.ds.orgs.OrgsDao">
+  <bean id="orgsDao" class="org.georchestra.ds.orgs.OrgsDaoImpl">
     <property name="ldapTemplate" ref="ldapTemplate"/>
     <property name="basePath" value="${ldapBaseDn}"/>
     <property name="orgTypeValues" value="${orgTypeValues:Association,Company,NGO,Individual,Other}"/>
